508 Compliance for collections/georef/batchgeoreftool.php (#962)
* 508 Compliance for collections/georef/batchgeoreftool.php

* add null coalescers to trim calls

* remove deprecated attributes from table and fix spacing and trim null coalescing in collections/georef/batchgeoreftool.php

* Fixed The Table

I fixed the table according to the information you provided on slack. But I am seeing a little issue with the Longitude (W  = input) where the "=" is not in place.

* Fixed Table and aria-label

* add skip nav link

* Implementation Fix

Implemented a reversal of the order for the Locality term and Generate List functionalities. Attempted to address the issue using the tabindex attribute, but found that while navigating through tabs, the Locality term was placed before the Home button.

---------

Co-authored-by: Mark <mark.aaron.fisher@gmail.com>