Closes #2276 Fixes dangling reference to photographer (#2282)
* Fix non audio media rendering as audio (#2280)

* Fix non audio media rendering as audio

# Issue Closes #2278

# Summary
Changes audio rendering to only occur if type is explicitly
MediaType::Audio
Image rendering will occur if audio failed the above new audio check and
is MediaType::Image or has a thumbnailUrl Present.

Add a failure else case that will render that the type is unknown and
could not be rendered

* Add Language tags for failed media render message

* Closes #2276 Fixes dangling reference to photographer

# Issue Closes #2276

# Summary
Uses 'creator' key instead of 'photographer' key. Was not caught because
it was behind spp permissions