Adds Support for Custom GeoJSON Overlays (#2517)
* Adds Support for Custom GeoJSON Overlays

# Issue Closes #2516

# Summary
Allows users to configure GeoJSON Overlays that are toggable on the map
search page.

Currently Supports custom popup templates although support is minimal
and this feature may not be desired.

All config is currently managed in symbini and with root server access
to upload the desired GeoJSON files. See template for configuration
instructions

* Add .gitignore to GeoJSON directory

* remove debug log

* Move geoJSON files location to content/geoJSON

* Update comment to show intent for map since list was confusing

* Add function to mount styles if present in the geoJSON file