# Closes 2541 Fixes issue with sortSequence or sortsequence not popul… (#2542)
* # Closes 2541 Fixes issue with sortSequence or sortsequence not populating when adding media

# Issue Closes #2541
# Summary
Because of the history of symbiota their is a schism in schema values
with sortsequence being used in older portals and sortSequence being
used in others. Some of that general confusion has bled into the logic.

To fix that some null coalencence to the correct value is needed. The
key is being stored as the lower case version `sortsequence` but because
of case insensitivity this will not be an issue at the insert level

* Add Both sortSequences to Media::update to cover cases where it isn't named consistently

* Fix getter to allow for both options

* create constant used for all media item query selection

* fix sort sequence on tpeditor add media

* fix casing on imgDetail update form

* Make update only used camel case keys and add it to the function documentation

* fix casing for hasValue, hasFunction and referenceUrl. archiveUrl is fixed in a parallel pr